+++ hata/discord/http/rate_limit_groups.py
@@ -860,6 +860,46 @@ Group Details
     - Limit : `OPT`
     - Resets after : `OPT`
 
+- auto_moderation_rule_get_all
+    - Endpoint : `/guilds/{guild_id}/auto-moderation/rules`
+    - Method: `GET`
+    - Required auth : `bot`
+    - Limiter : `guild_id`
+    - Limit : `UN`
+    - Resets after : `UN`
+
+- auto_moderation_rule_create
+    - Endpoint : `/guilds/{guild_id}/auto-moderation/rules`
+    - Method: `POST`
+    - Required auth : `bot`
+    - Limiter : `guild_id`
+    - Limit : `UN`
+    - Resets after : `UN`
+
+- auto_moderation_rule_get
+    - Endpoint : `/guilds/{guild_id}/auto-moderation/rules/{auto_moderation_rule_id}`
+    - Method: `GET`
+    - Required auth : `bot`
+    - Limiter : `guild_id`
+    - Limit : `UN`
+    - Resets after : `UN`
+
+- auto_moderation_rule_delete
+    - Endpoint : `/guilds/{guild_id}/auto-moderation/rules/{auto_moderation_rule_id}`
+    - Method: `DELETE`
+    - Required auth : `bot`
+    - Limiter : `guild_id`
+    - Limit : `UN`
+    - Resets after : `UN`
+
+- auto_moderation_rule_edit
+    - Endpoint : `/guilds/{guild_id}/auto-moderation/rules/{auto_moderation_rule_id}`
+    - Method: `PATCH`
+    - Required auth : `bot`
+    - Limiter : `guild_id`
+    - Limit : `UN`
+    - Resets after : `UN`
+
 - guild_ban_get_chunk
     - Endpoint : `/guilds/{guild_id}/bans`
     - Method : `GET`
@@ -1959,6 +1999,11 @@ guild_get = RateLimitGroup(LIMITER_GUILD, optimistic=True)
 guild_edit = RateLimitGroup(LIMITER_GUILD, optimistic=True)
 guild_ack = RateLimitGroup() # untested
 audit_log_get_chunk = RateLimitGroup(LIMITER_GUILD, optimistic=True)
+auto_moderation_rule_get_all = RateLimitGroup(LIMITER_GUILD, optimistic=True)
+auto_moderation_rule_create = RateLimitGroup(LIMITER_GUILD, optimistic=True)
+auto_moderation_rule_get = RateLimitGroup(LIMITER_GUILD, optimistic=True)
+auto_moderation_rule_delete = RateLimitGroup(LIMITER_GUILD, optimistic=True)
+auto_moderation_rule_edit = RateLimitGroup(LIMITER_GUILD, optimistic=True)
 guild_ban_get_chunk = RateLimitGroup(LIMITER_GUILD)
 guild_ban_delete = RateLimitGroup(LIMITER_GUILD, optimistic=True)
 guild_ban_get = RateLimitGroup(LIMITER_GUILD, optimistic=True)